HeadGAN/requirements.txt
-f https://download.pytorch.org/whl/cu113
gdown
torch == 1.11.0
torchvision == 0.12.0
# 3DDFA_V2
matplotlib
numpy
opencv-python # cv2
imageio
imageio-ffmpeg
pyyaml # yaml
tqdm
argparse
cython
scikit-image # skimage
scipy
onnxruntime
gradio
#pyAudioAnalysis
matplotlib==3.4.2
simplejson==3.16.0
scipy==1.6.3
numpy==1.20.3
hmmlearn==0.2.5
eyeD3==0.9.6
pydub==0.25.1
scikit_learn==0.24.2
tqdm==4.52.0
plotly==5.3.1
pandas==1.2.4
imblearn
tensorboard
tabulate
